refactor(structure): reorganize project structure
