+++ src/configs/browser.config
@@ -22,7 +22,7 @@
     },
     {
         "name": "Brave",
-        "icon": { "css": null, "localPath": "./images/browserIcons/Brave.svg" },
+        "icon": { "css": null, "localPath": "./../../assets/browserIcons/Brave.svg" },
         "scriptLocation": "interScript",
         "startScript": "proxy-brave-on.sh",
         "stopScript": "proxy-brave-off.sh"
