+++ src/renderer/main/style.css
