+++ src/renderer/settings/settings.html
@@ -117,10 +117,9 @@
                     <div class="script-hint" id="script-hint">
                         If the new browser is Firefox-based or Chromium-based,
                         use the corresponding scripts from the default configuration
-                        as a reference.
-                        <a href="#" id="open-scripts-folder">
-                            /opt/psiphonlinuxgui/resources/app/scripts/
-                        </a>
+                        as a reference. You can find them in the
+                        <a href="https://github.com/Alessandros-Hube/PsiphonLinuxGUI/tree/main/scripts"
+                            target="_blank">scripts</a> folder of the project repository.
                     </div>
 
                     <div class="modal-actions">
